refactor: Inline validation code to make changes easier
