+++ src/shared/lib/PasswordValidationResult.ts
@@ -1,0 +1,6 @@
+export enum PasswordValidationResult {
+  AT_LEAST_8,
+  MAXIMUM_72,
+  COMMON_PASSWORD,
+  VALID_PASSWORD,
+}
